de_config_tool_modified/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.4
httpx>=0.27.0
pydantic>=2.0.0

# AWS
boto3>=1.34.0

# Azure
azure-storage-file-datalake>=12.15.0
azure-eventhub>=5.12.0
azure-identity>=1.16.0
# azure-mgmt-datafactory>=6.0.0  # Uncomment for real ADF extraction

# Databricks
# databricks-sdk>=0.28.0         # Uncomment for real Databricks extraction

# Snowflake
snowflake-connector-python>=3.9.0

# Databases / Sources
asyncpg>=0.29.0
aiomysql>=0.2.0
aioodbc>=0.5.0
aiokafka>=0.10.0
motor>=3.4.0
simple-salesforce>=1.12.5
paramiko>=3.4.0
google-cloud-storage>=2.16.0
